// dashboard/src/api/models/nodeResultDeltaEvent.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Scheduler Public API (v1)
 * OpenAPI spec version: 1.3.0
 */
import type { NodeResultDeltaEventOperation } from './nodeResultDeltaEventOperation';
import type { NodeResultDeltaEventPayload } from './nodeResultDeltaEventPayload';
import type { JsonPatchOperation } from './jsonPatchOperation';
import type { NodeResultDeltaEventChunkMeta } from './nodeResultDeltaEventChunkMeta';

export interface NodeResultDeltaEvent {
  kind: 'node.result.delta';
  runId: string;
  nodeId: string;
  /** @minimum 0 */
  revision: number;
  /** @minimum 0 */
  sequence: number;
  operation: NodeResultDeltaEventOperation;
  /**
   * @nullable
   * @pattern ^(/[^/]+)*$
   */
  path?: string | null;
  /** @nullable */
  payload?: NodeResultDeltaEventPayload;
  /** @nullable */
  patches?: JsonPatchOperation[] | null;
  /** @nullable */
  chunkMeta?: NodeResultDeltaEventChunkMeta;
  terminal?: boolean;
}