// dashboard/src/api/models/runRef.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Scheduler Public API (v1)
 * OpenAPI spec version: 1.3.0
 */
import type { RunStatus } from './runStatus';

export interface RunRef {
  runId: string;
  status: RunStatus;
  definitionHash: string;
  clientId: string;
  createdAt: string;
}